+++ ShoutKitApp/ShoutKitWidgets/ShoutKitWidgetsBundle.swift
@@ -5,5 +5,6 @@ import WidgetKit
 struct ShoutKitWidgetsBundle: WidgetBundle {
     var body: some Widget {
         NowPlayingLiveActivity()
+        QuickPlayWidget()
     }
 }